Simplify module composition
This commit simplifies the composition of modules.

The root module configures a function sourced from a directory on
localhost. The function-sourced-from-repository submodule offers an
alternative implementation which configures a function sourced from a
repository.

Events are composed as singular submodules to strike a balance between
extensibility and ease of use.