fix: remove gzip content-encoding metadata on function source archive
The gzip content encoding is not right for a zip archive and does not
actually work. When working with this module, I had to remove the
content-encoding to make things work. Otherwise, I was getting a
"End-of-central-directory signature not found" error in Cloud Build
during function deployment.

When I tried to download the archive using my browser, the browser
errored, but gsutil succeeded. The file uploaded to GCS was complete so
that was not the reason for the error. Removing the content-encoding
field fixed the deployment.